This is a simple move of checks/checks.go to healthcheck/wait.go
partial #383
